Refactor: Extract applyAliases logic into shared extension function
