+++ app/src/main/java/io/pm/finlight/ui/viewmodel/TransactionViewModel.kt
@@ -22,6 +22,7 @@ import io.pm.finlight.ui.viewmodel.AnalysisTransactionType
 import io.pm.finlight.utils.CategoryIconHelper
 import io.pm.finlight.utils.FormatUtils
 import io.pm.finlight.utils.HeuristicCategorizer
+import io.pm.finlight.utils.applyAliases
 import io.pm.finlight.utils.ShareImageGenerator
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.ExperimentalCoroutinesApi
@@ -402,7 +403,7 @@ class TransactionViewModel(
                         emit(emptyList())
                     }
             }.combine(merchantAliases) { transactions, aliases ->
-                applyAliases(transactions, aliases)
+                transactions.applyAliases(aliases)
             }.stateIn(viewModelScope, SharingStarted.WhileSubscribed(5000), emptyList())
 
         val financialSummaryFlow =
@@ -906,42 +907,10 @@ class TransactionViewModel(
         }
     }
 
-    private fun applyAliases(
-        transactions: List<TransactionDetails>,
-        aliases: Map<String, String>,
-    ): List<TransactionDetails> {
-        return transactions.map { details ->
-            val original = details.transaction.originalDescription
-            val currentDesc = details.transaction.description
-            val key = (original ?: currentDesc).lowercase(Locale.getDefault())
-            val alias = aliases[key]
-
-            val newDescription =
-                if (alias != null) {
-                    // Only apply the alias if the current description matches the original description.
-                    // If it does not match, it means the user manually changed it to something else,
-                    // or it was already evaluated to something else.
-                    if (original != null && currentDesc.equals(original, ignoreCase = true)) {
-                        alias
-                    } else if (currentDesc.equals(alias, ignoreCase = true)) {
-                        alias
-                    } else if (original == null && currentDesc.equals(key, ignoreCase = true)) {
-                        // If there's no original description, and the current description is the key, apply alias.
-                        alias
-                    } else {
-                        currentDesc
-                    }
-                } else {
-                    currentDesc
-                }
-            details.copy(transaction = details.transaction.copy(description = newDescription))
-        }
-    }
-
     fun findTransactionDetailsById(id: Int): Flow<TransactionDetails?> {
         return transactionRepository.getTransactionDetailsById(id)
             .combine(merchantAliases) { details, aliases ->
-                details?.let { applyAliases(listOf(it), aliases).firstOrNull() }
+                details?.let { listOf(it).applyAliases(aliases).firstOrNull() }
             }
     }
